added mail service

+++ api/routes/auth.js
@@ -2,6 +2,7 @@ const router = require("express").Router();
 const User = require("../models/User")
 const OTP = require("../models/Otp")
 const bcrypt = require("bcrypt");
+const { sendMail } = require("../mailer/mailer");
 
 // register user
 router.post('/register', async (req, res) => {
@@ -59,12 +60,17 @@ router.get('/email/:email', async (req, res) => {
     try {
         const oldUser = await User.findOne({ email: req.params.email });
         if (!oldUser) {
-            const otp = Math.floor((Math.random() * 1000000) + 1);
+            let otp = 0;
+            while (otp < 100000) otp = Math.floor((Math.random() * 1000000) + 1);
             const newOTP = new OTP({
                 email: req.params.email,
                 otp,
             })
             const savedOtp = await newOTP.save();
+            sendMail({
+                userEmail: req.params.email,
+                code: otp,
+            })
             console.log(savedOtp);
             return res.status(200).json({ msg: "all ok", otpId: savedOtp._id });
         } else {
@@ -81,8 +87,9 @@ router.post('/otp', async (req, res) => {
     try {
         const oldOTP = await OTP.findById(req.body.otpId);
         if (!oldOTP) return res.status(404).json({ msg: "otp expired" });
-        if (oldOTP.otp === req.body.otp) {
+        if (oldOTP.otp === req.body.otp && oldOTP.email === req.body.email) {
             await oldOTP.deleteOne();
+            await OTP.deleteMany({ email: req.body.email });
             return res.status(200).json({ msg: "success" })
         }
         return res.status(403).json({ msg: "otp mismatch" });
